eduard_backend-master/requirements.txt
Django==4.2.3
django-allauth==0.54.0
django-axes==6.1.0
bmi-topography==0.8.2
djangorestframework==3.14.0
numpy==1.25.2
scipy==1.11.2
opencv-python==4.8.0.76
matplotlib==3.7.2
google-auth==2.22.0
python-dotenv==1.0.0
django-rest-knox==4.2.0
torchvision==0.14.0
torch==1.13.0
torchaudio==0.13.0
Pillow~=10.0.0
rasterio~=1.3.8
django-cors-headers==4.2.0
django-storages==1.14
stripe==6.6.0